pass pointer to meddlesome function
Illustrate passing a pointer to a function. Within the function, the formal parameter is a *copy* of
the pointer. That is, it is another pointer, pointing at the same location. Thus, changing where it
points does not affect the original pointer, but changing the data *to which* it points does just
that.